Normalizing sound output

+++ Sound.h
@@ -14,5 +14,5 @@ public:
   }
 
   // This is uint16 only because of how the mixer is implemented ... need to look into a more generic method
-  virtual uint8 getSamplesForFrame(int16* samples, uint8 channels, uint32 frame) = 0;
+  virtual uint8 getSamplesForFrame(float* samples, uint8 channels, uint32 frame) = 0;
 };
